Test fix 264768 for 259993 : On WebSphere 7.0 we hang on a find() after a commit() http://bugs.eclipse.org/264768 Reviewed by: Andrei
I discussed this with Andrei and code reviewed it.
The test case must be adapted to check that the entity that was cached in the
closed entityManager is still managed or not.
    public void testGetReferenceUsedInUpdate()
    public void testCollectionAddNewObjectUpdate() 

Some application servers like WebSphere 7.0 will em.clear() the entityManager during a commit() before releasing the em back to their em pool for GC