Fix for issue 283092.
Accessing Oracle driver classes eagerly to fail fast when Oracle driver classes are not available to EclipseLink  while running against Oracle DB. This would be the case when some one is using non Oracle driver or running in an environment where the driver classes are not visible to EclipseLink (e.g. running against GlassGish when driver jars are put in lib dir)

The eager failure will trigger code in DataSourceLogin to fall back to OraclePlatform. 

Code reviewed by Tom.