For for Bug 299559 - JAXB MARSHALS WITH EMPTY TAG NAME WHEN NOT USING JAXBELEMENT
Background Information:
The possible root elements for a descriptor are stored in a list, the first one is the default, the others are secondary ones.  At some point we introduced the possibility of having secondary root elements without a primary one.  To implement this we specify the default root element as empty string ("").

Core Fix:
When a default root element is being set on an XMLDescriptor if the root element name is "", then this will not result in default root element field with a "" XPath.

JAXB Fix:
The following are now the order in where the root element comes from when using a TypeMappingInfo in the marshal operation:
1.  If the object is wrapped in a JAXBElement then the element name from the JAXBElement will be used.
2.  If no JAXB element, and a tagName is specified on the TypeMappingInfo then the tagName property will be used.
3.  If no JAXB element, and no tagName is specified on the TypeMappingInfo then if a default root element is specified on the descriptor, then the default root element will be used.
4.  If no JAXB element, and no tagName is specified on the TypeMappingInfo, and the descriptor has no default root element, then no root element will be written out.

Code Review By:  Denise Smith

Checked into 2.0.1 at rev: 6282