Bug 300556 - Can't configure descriptor to make sequence override existing negative pk values
Suggested patch for 2.0.1

The patch was suggested by James.

IdValidation is extended, the new NEGATIVE value allows to specify that
negative and zero values are not allowed in primary key;

The same method isPrimaryKeyComponentInvalid in ObjectBuilder is now used for
both validation of the extracted primary key and deciding whether the existing
pk value should be overridden by sequencing;

A new method shouldAlwaysOverrideExistingValue was defined in Sequence class to
indicate whether the existing pk value should be overridden by sequence even if
it's a valid pk value (by default only Identity does that).

A new variable defaultIdValidation was defined in Project class that is used as
a default value for ClassDescriptor.idValidation.
To make Eclipselink override nulls, negative and zero values it should be set
to IdValidation.NEGATIVE before session login.

A new PersistenceUnitProperty "eclipselink.id-validation" was defined to access
it in JPA.
To make Eclipselink override nulls, negative and zero values this property
should be set to "NEGATIVE" either in persistence.xml or passed to
createEntityManagerFactory.

In core tests RegisterNewObjectInIdentityMapNoSeqTest was expanded to test
different idValidations with and without sequencing.

In jpa tests EntityManagerJUnitTestSuite.testPersistenceProperties was expanded
to test for the new property and it being overridden by @PrimaryKey annotation.