Fix for Bug 307939 -  Add support for ObjectReferenceMapping without a reference descriptor
We now support null or Object as the reference class on XMLObjectReferenceMapping and XMLCollectionReferenceMapping (see below).  Note that when no reference class is set, the mapping is limited to one pk/fk pair.

     XMLCollectionReferenceMapping phoneNumbersMapping = new XMLCollectionReferenceMapping();
     phoneNumbersMapping.setAttributeName("phoneNumbers");
     phoneNumbersMapping.addSourceToTargetKeyFieldAssociation("phone-number-id/text()", null);
     phoneNumbersMapping.setReferenceClass(Object.class);
     phoneNumbersMapping.getContainerPolicy().setContainerClass(ArrayList.class);
     xmlDescriptor.addMapping(phoneNumbersMapping);

With this type of mapping when the reference is being resolved each cache with a primary key size of 1 will be checked.

Code Reviewed By:  Rick Barkhouse