Bug 305713 - OuterJoinExpressionHolder linearization order can be incorrect during ansi joins
Updated SQLSelectStatement to accommodate James' changes (outer join for map
key).
Had to alter a couple of test classes: 
JoinedAttributeTestHelper - to take into account that for OneToOne mapping in
case the foreign key is null the value holder in the read object is always
triggered;
ReportQueryMultipleReturnTestSuite.testReturnRootObject - had to change the
reference class, see the comment in the test.