Bug 323880 - "This is already a proxy session" exception on WLS 10.3.3 after explicitly rolling back the user transaction patch with two disabled tests
The core problem is in WLS and it will be fixed in WLS 10.3.4.
The current patch
1) Adds debug output to connection customizer;
2) Adds a check for already opened proxy session - that should work around some
(but not all!) problems: the connection released into poll without clearing its
proxy session will have its old proxy session closed before starting a new one.
That should fix the problem reported by the bug filer (see
ProxyAuthenticationServerTestSuite.testFlushRollback);
3) Added two commented out non-Eclipselink tests: testJtaDataSource and
testNonJtaDataSource that illustrate the core (WLS) problem. The tests should
be uncommented when the core bug is fixed in WLS 10.3.4.

All what's left here is to wait until the core WLS bug is fixed, then uncomment
the tests and close the bug.