Fix for Bug 333306 - MOXy: Issues with Self Mappings and XPaths to same name
Code reviewed by:  David McCann

Normally during an unmarshal when unmapped content is found, an unmapped content handler is set on the XMLReader.  When the marshal record has self records this can not occur.  In these cases a flag has been added to UnmarshalRecord to keep track of being in an unmapped state, this is done by level.