Bug 328404 - JPA Persistence Unit Composition Added support for connection properties for composite member persistence unit.
In the same way as PersistenceUnitProperties.COMPOSITE_UNIT_PROPERTIES used to
pass member-specific properties to createEMF method now is possible to pass
connection properties (to be processed by members) to createEM method.
Tests are in composite.advanced.EntityManagerJUnitTestSuite:
testConnectionPolicy;
testConnectionPolicySetProperty;
testNonPooledConnection;
testExclusiveIsolatedLeaksConnectionOnClear.

Note that lots of formatting changes in
composite.advanced.EntityManagerJUnitTestSuite were done to make as close as
possible to the original class advanced.EntityManagerJUnitTestSuite

Ideally all the tests found in the latter should be copied to the former (some
adaptation may be required).