Bug 328404 - JPA Persistence Unit Composition Patch adds @CompositeMember annotation and xml support.
A CompositeMember annotation is ignored unless is in composite member
persistence unit.  
It may be used in conjunction with a ElementCollection and CollectionTable
annotations.  
It should be used if target type is a primitive type and CollectionTable
designates the table that belongs to composite member persistence unit other
than the source composite member persistence unit.
That allows the source and target to be mapped to different databases.