Bug 350487 - JPA 2.1 Specification defined support for Stored Procedure Calls
Verified by: Peter Krogh, Eric Gwin, Tom Ware

Notes: This is a development stamp of this feature. It also includes Eclipse project changes and updates as well as new JPA 2.1 test launch suite and various ant script changes for building.

Tests: A new JPA 2.1 TestSuite launch is provided and contains tests of the new stored procedures (testing what is included in this patch for this feature thus far)