Bug #431803: MOXy support for Bean Validation (JSR-303/349)
Fix for endless invocation loop between
 unmarshaller - validator - unmarshaller.

Java docs coverage increased to 100 %.

META-INF/validation.xml is the test for this bug-fix. In order to run this test, developer still needs to force Bean Validation provider (e.g. HV) to use MOXy. That cannot be committed since some of our tests depend on JAXB-RI as default JAXB provider. The solution I used was to modify HV and repackage it.

Signed-off-by: Marcel Valovy <marcel.valovy@oracle.com>