Fix for #449661.
All tests pass.
Commit exactly reflects what was reviewed and approved by Martin Vojtek on orahub.

Problem that this fixes:
IntelliJ IDEA 13 managed to commit historical files that corresponded to file names instead
of correct files that were present in the directory because of problem with module settings.

The solution was to copy the files committed to orahub to the workspace directory and remove
the old files and prepare this commit without turning IntelliJ IDEA on.

Signed-off-by: Marcel Valovy <marcel.valovy@oracle.com>