Avoid ClassCastException and IllegalArgumentException in cloneFor implementations
VectorContainerPolicy is used for all List types.  Likely the Object
passed to cloneFor will not be a Vector, but rather an ArrayList,
LinkedList or some other List derivative.  Getting a ClassCastException
is expected to be the normal code path so doing an instanceof check to
avoid the exception.  instanceof will perform better than the
ClassCastException since you need to do a stack walk in order to
generate the stack trace of the exception.  In general Exceptions should
not be a normal decision mechanism if possible.

Signed-off-by: Jared Anderson <jhanders@us.ibm.com>