Bug 417725 - MOXyJsonProvider - Support writing collection/array of datatypes
This is also fix for Bug 459227 - Moxy is not able to marshall List<String> into JSON.
Moxy now is able to marshall arrays and collections of primitives and primitive wrappers into JSON when as an input object is only this array or collection.
These data types are marked as a writable in MOXyJsonProvider.isWriteable .
Unnecessary commented line in test method removal.

See https://bugs.eclipse.org/bugs/show_bug.cgi?id=417725
    https://bugs.eclipse.org/bugs/show_bug.cgi?id=459227

Signed-off-by: Radek Felcman <radek.felcman@oracle.com>