[2.7] JPA test fix for SELECT...FOR UPDATE WAIT xx
This is fix for JPA with `SELECT...FOR UPDATE WAIT xx` queries.
It's executed on OraclePlatform only.
Some description for SQL SELECT FOR UPDATE WAIT xx option (from Oracle DB doc):

The NOWAIT and WAIT clauses let you tell the database how to proceed if
the SELECT statement attempts to lock a row that is locked by another user.

Specify WAIT to instruct the database to wait integer seconds for the row to become available and then return control to you.

Before this fix `QueryHints.PESSIMISTIC_LOCK_TIMEOUT` was applied to first query, but
to test this feature correctly it should be applied to the second query in test.
This is not instruction how long DB keeps lock, but how long query will
wait for a lock release if there are some locked rows in the selection.
To simulate real environment more more accurately some timeout and lock release
instructions for a first query are executed in parallel (in thread) with second query.

`testPESSIMISTIC_LockWithDefaultTimeOutUnit` expects, that selected rows are
available before timeout (FOR UPDATE WAIT xx)

`testPESSIMISTIC_LockWithSecondsTimeOutUnit` expects `javax.persistence.LockTimeoutException`
Selected rows will not be available before timeout (FOR UPDATE WAIT xx)

https://docs.oracle.com/en/database/oracle/oracle-database/12.2/sqlrf/SELECT.html#GUID-CFA006CA-6FF1-4972-821E-6996142A51C6
Signed-off-by: Radek Felcman <radek.felcman@oracle.com>