Various fixes for using "this" variable in JPQL (#2224)
* Fixes #2197, #2198, #2199 related to using "this" variable in JPQL

When the "this" variable is used explicitly, it's virtual but needs to be skipped
in expressions that expect a virtual variable.
The problem was that expressions like 'this.field' were treated as
if 'this' was a field on the implicit variable.

* Adds a test for #2183

The issue is already fixed. This is only a test that verifies the fix.