Clone batchQuery when executing to ensure nested batch fetches retain proper rows (#2386)
Closes #1998 
This PR changes the execution of the batchQuery to utilize a clone for actual execution of the batch fetching. Without the clone it was possible that further batches would overwrite the DataResults of the original batchQuery. These DataResults were then used by the various BatchValueHolders to instantiate the values lazily. However when overwritten the referenced rows had the possibility of not existing in the policy as only the most recently executed batch was present. This could lead to incorrect results being returned.