Bind partial parameters for DB2
shouldBindPartialParameters was set to true in DB2Platform, but later
overridden by a copy of the default parameters from the default
DBPlatform.

It never actually worked.

After enabling it for real, found out that for
"translateQueryStringAndBindParameters" in DatabaseCall,
"bindParameters" (for that call) has to be checked for, otherwise things
like setting query hints to not bind from the caller site don't work
anymore.

Additionally, "dynamicSQLRequiredForFunctions" should return false when
"shouldBindPartialParameters" is true. Otherwise a single operator will
immediately disable binding for the entire query, which is the exact
opposite of what partial binding is about.

A large number of tests had to be adjusted for the new output of DB2.